… internal/adapter/proxy/benchmark_comparison_test.go
… internal/adapter/proxy/benchmark_refactor_test.go
… internal/adapter/proxy/config_olla.go
… internal/adapter/proxy/core/common_test.go
… internal/adapter/proxy/proxy_benchmark_test.go
… internal/adapter/proxy/proxy_headers_test.go
… internal/adapter/proxy/proxy_integration_test.go
… internal/adapter/proxy/proxy_olla.go
… internal/adapter/proxy/proxy_olla_test.go
… internal/adapter/proxy/proxy_path_stripping_test.go
… internal/adapter/proxy/proxy_routing_test.go
… internal/adapter/proxy/proxy_sherpa.go
… internal/adapter/proxy/proxy_sherpa_test.go
… internal/adapter/proxy/proxy_test.go
… internal/adapter/proxy/sherpa/service.go
… internal/adapter/proxy/sherpa/service_leak_stress_test.go
… internal/adapter/proxy/sherpa/service_leak_test.go
… internal/adapter/proxy/sherpa/service_streaming.go
… internal/app/config.go
… internal/app/handlers/handler_common.go
… internal/app/handlers/handler_provider_common.go
… internal/app/handlers/handler_proxy.go
… internal/core/constants/context.go
… internal/logger/styled_plain.go
… internal/logger/styled_pretty.go
… internal/router/registry.go
… makefile
… pkg/eventbus/eventbus.go
… pkg/eventbus/eventbus_stress_test.go
… pkg/eventbus/eventbus_test.go