This PR tackles consolidating the two proxies and avoiding the excessive duplication across the two now that they both are considered stable implementation wise. Also fixes a few annoying bugs and improves performance.

- Olla and Sherpa now share the same configuration structs.
- Olla and Sherpa proxies now reuse a lot more code (but maintain their traits)
- BUGFIX: Path matching with StripPrefix(`/olla` prefix will incorrectly strip `/ollama` → `/ma` across patterns sometimes)
- BREAK: We used to  always send `X-Forwarded-Host` (like in scout/sherpa) but now we only send if it's not present, this may break for multi-hop proxies.

However, we looked at tuning it even further based on the internal benchmark.

## Summary of changes

| Benchmark                       | Metric      | Baseline | Current | Δ         | Change                          |
| ------------------------------- | ----------- | -------- | ------- | --------- | ------------------------------- |
| **ErrorHandling**               | `ns/op`     | 28,126   | 5,972   | 🟢 -78.8% | 4.7× faster                     |
|                                 | `allocs/op` | 154      | 83      | 🟢 -46%   | Fewer allocations               |
|                                 | `B/op`      | 12,241   | 7,233   | 🟢 -40.9% | Less memory                     |
| **ConfigUpdates**               | `ns/op`     | 49.43    | 57.54   | 🟡 +16.4% | Slight trade-off                |
| **StatsCollection**             | `ns/op`     | 35.00    | 8.17    | 🟢 -76.6% | 4.3× faster                     |
| **HeaderProcessing**            | `ns/op`     | 111,672  | 98,838  | 🟢 -11.5% | Faster processing               |
|                                 | `allocs/op` | 257      | 224     | 🟢 -12.8% | Fewer allocations               |
|                                 | `B/op`      | 22,618   | 16,899  | 🟢 -25.3% | Less memory                     |
| **GetStats (Sherpa)**           | `ns/op`     | 35.59    | 7.75    | 🟢 -78.2% | 4.6× faster                     |
| **GetStats (Olla)**             | `ns/op`     | 34.94    | 7.72    | 🟢 -77.9% | 4.5× faster                     |
| **ErrorFunction**               | `ns/op`     | 475.5    | 536.5   | 🔴 +12.8% | Minor regression                |
| **MemoryUsage (Sherpa)**        | `ns/op`     | 95,751   | 104,199 | 🔴 +8.8%  | Slight increase                 |
| **MemoryUsage (Olla)**          | `ns/op`     | 94,634   | 98,746  | 🔴 +4.3%  | Acceptable                      |
|                                 | `allocs/op` | 209      | 187     | 🟢 -10.5% | Improved                        |
| **ConnectionReuse (Sherpa)**    | `ns/op`     | 87,066   | 69,172  | 🟢 -20.6% | Faster reuse                    |
| **ConnectionReuse (Olla)**      | `ns/op`     | 25,401   | 10,517  | 🟢 -58.6% | 2.4× faster                     |
|                                 | `allocs/op` | 208      | 185     | 🟢 -11%   | Improved                        |
| **CircuitBreaker (Check)**      | `ns/op`     | 3.20     | 16.64   | 🔴 +420%  | New cost from granular tracking |
| **CircuitBreaker (Concurrent)** | `ns/op`     | —        | 15.84   | ⚪ —       | We  added this                  |
| **CircuitBreaker (Threshold)**  | `ns/op`     | —        | 143.7   | ⚪ —       | We  added this                  |


Olla's refactor reduced latency and allocations across most benchmarks (up to 4.7× faster), with minor regressions in circuit breaker cost due to added tracking.

## Baseline

`main` commit [d03fe4...d94](https://github.com/thushan/olla/commit/d03fe40a7bbd251d9ef92c60ef59079b9d243d94)

```
BenchmarkProxyImplementations/Olla/ErrorHandling-16           	   43749	     28126 ns/op	   12241 B/op	     154 allocs/op
BenchmarkProxyImplementations/Olla/ConfigUpdates
BenchmarkProxyImplementations/Olla/ConfigUpdates-16           	23395597	        49.43 ns/op	      80 B/op	       1 allocs/op
BenchmarkProxyImplementations/Olla/StatsCollection
BenchmarkProxyImplementations/Olla/StatsCollection-16         	33153475	        35.00 ns/op	       0 B/op	       0 allocs/op
BenchmarkProxyImplementations/Olla/HeaderProcessing
BenchmarkProxyImplementations/Olla/HeaderProcessing-16        	   10795	    111672 ns/op	   22618 B/op	     257 allocs/op
BenchmarkProxyStats
BenchmarkProxyStats/Sherpa_GetStats
BenchmarkProxyStats/Sherpa_GetStats-16                        	32711808	        35.59 ns/op	       0 B/op	       0 allocs/op
BenchmarkProxyStats/Olla_GetStats
BenchmarkProxyStats/Olla_GetStats-16                          	33994333	        34.94 ns/op	       0 B/op	       0 allocs/op
BenchmarkErrorFunction
BenchmarkErrorFunction-16                                     	 2502780	       475.5 ns/op	     137 B/op	       4 allocs/op
BenchmarkMemoryUsage
BenchmarkMemoryUsage/Sherpa_MemoryAllocations
BenchmarkMemoryUsage/Sherpa_MemoryAllocations-16              	   12224	     95751 ns/op	   18823 B/op	     210 allocs/op
BenchmarkMemoryUsage/Olla_MemoryAllocations
BenchmarkMemoryUsage/Olla_MemoryAllocations-16                	   12624	     94634 ns/op	   18576 B/op	     209 allocs/op
BenchmarkConnectionPooling
BenchmarkConnectionPooling/Sherpa_ConnectionReuse
BenchmarkConnectionPooling/Sherpa_ConnectionReuse-16          	   16390	     87066 ns/op	   31445 B/op	     224 allocs/op
BenchmarkConnectionPooling/Olla_ConnectionReuse
BenchmarkConnectionPooling/Olla_ConnectionReuse-16            	   46867	     25401 ns/op	   20269 B/op	     208 allocs/op
BenchmarkCircuitBreaker
BenchmarkCircuitBreaker/Olla_CircuitBreakerCheck
BenchmarkCircuitBreaker/Olla_CircuitBreakerCheck-16           	362502390	         3.200 ns/op	       0 B/op	       0 allocs/op
BenchmarkObjectPools
BenchmarkObjectPools/Olla_BufferPool
BenchmarkObjectPools/Olla_BufferPool-16                       	100000000	        12.16 ns/op	       0 B/op	       0 allocs/op
BenchmarkObjectPools/Olla_RequestContextPool
BenchmarkObjectPools/Olla_RequestContextPool-16               	100000000	        11.84 ns/op	       0 B/op	       0 allocs/op
PASS
```

## Consolidated & Optimised Further

This includes new rounds of profiling and heavy traffic analysis on our busier nodes.

```
BenchmarkProxyImplementations/Olla/ErrorHandling-16           	  281624	      5972 ns/op	    7233 B/op	      83 allocs/op
BenchmarkProxyImplementations/Olla/ConfigUpdates
BenchmarkProxyImplementations/Olla/ConfigUpdates-16           	35091412	        57.54 ns/op	      80 B/op	       1 allocs/op
BenchmarkProxyImplementations/Olla/StatsCollection
BenchmarkProxyImplementations/Olla/StatsCollection-16         	145893889	         8.171 ns/op	       0 B/op	       0 allocs/op
BenchmarkProxyImplementations/Olla/HeaderProcessing
BenchmarkProxyImplementations/Olla/HeaderProcessing-16        	   12396	     98838 ns/op	   16899 B/op	     224 allocs/op
BenchmarkProxyStats
BenchmarkProxyStats/Sherpa_GetStats
BenchmarkProxyStats/Sherpa_GetStats-16                        	153806734	         7.750 ns/op	       0 B/op	       0 allocs/op
BenchmarkProxyStats/Olla_GetStats
BenchmarkProxyStats/Olla_GetStats-16                          	154590666	         7.715 ns/op	       0 B/op	       0 allocs/op
BenchmarkErrorFunction
BenchmarkErrorFunction-16                                     	 2212374	       536.5 ns/op	     137 B/op	       4 allocs/op
BenchmarkMemoryUsage
BenchmarkMemoryUsage/Sherpa_MemoryAllocations
BenchmarkMemoryUsage/Sherpa_MemoryAllocations-16              	   12654	    104199 ns/op	   16619 B/op	     206 allocs/op
BenchmarkMemoryUsage/Olla_MemoryAllocations
BenchmarkMemoryUsage/Olla_MemoryAllocations-16                	   12193	     98746 ns/op	   14418 B/op	     187 allocs/op
BenchmarkConnectionPooling
BenchmarkConnectionPooling/Sherpa_ConnectionReuse
BenchmarkConnectionPooling/Sherpa_ConnectionReuse-16          	   21009	     69172 ns/op	   18258 B/op	     213 allocs/op
BenchmarkConnectionPooling/Olla_ConnectionReuse
BenchmarkConnectionPooling/Olla_ConnectionReuse-16            	  141662	     10517 ns/op	   14411 B/op	     185 allocs/op
BenchmarkCircuitBreaker
BenchmarkCircuitBreaker/CircuitBreakerCheck
BenchmarkCircuitBreaker/CircuitBreakerCheck-16                	75242185	        16.64 ns/op	       0 B/op	       0 allocs/op
BenchmarkCircuitBreaker/CircuitBreakerFailureThreshold
BenchmarkCircuitBreaker/CircuitBreakerFailureThreshold-16     	 8753750	       143.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkCircuitBreaker/CircuitBreakerConcurrent
BenchmarkCircuitBreaker/CircuitBreakerConcurrent-16           	71275837	        15.84 ns/op	       0 B/op	       0 allocs/op
PASS
```


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

**New Features**
- Introduced new Sherpa and Olla proxy services with enhanced performance, streaming, error handling, and connection management tailored for AI inference workloads.
- Added extensive benchmarking suites comparing proxy implementations on request handling, streaming, configuration updates, and circuit breaker behavior.
- Implemented detailed, user-friendly error messages for proxy failures with actionable guidance.
- Added advanced proxy configuration options for connection pooling, timeouts, and TCP tuning.
- Introduced a new streaming response test script validating incremental streaming quality across multiple LLM providers.

**Bug Fixes**
- Improved and rigorously tested route prefix stripping and request path forwarding to prevent double-stripping and ensure accurate upstream paths.
- Enhanced HTTP header filtering to block sensitive headers and correctly add proxy metadata headers.

**Documentation**
- Expanded documentation with detailed instructions for the new streaming response test script and updated test running procedures.

**Refactor**
- Replaced previous proxy implementations with refactored, encapsulated Sherpa and Olla services exposing public constructors and methods.
- Unified proxy core components for event publishing, statistics tracking, and header management.
- Introduced a worker pool for event bus asynchronous publishing to improve concurrency and resource management.
- Updated proxy factory to support error-aware proxy creation and improved type clarity.

**Tests**
- Added comprehensive unit, integration, and leak detection tests covering proxy routing, streaming, error handling, circuit breakers, connection pools, and header management.
- Modernised existing proxy tests to use public APIs and constructors, removing direct internal state inspection.
- Added new tests for proxy path stripping, routing regressions, and streaming client disconnect handling.

**Chores**
- Renamed context key constants for improved clarity and consistency.
- Updated logger implementations to use centralized context key constants.
- Improved test scripts and load testing commands with timeouts and error handling enhancements.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->